moved test files around
