+++ bin/micromustache.spec.js
@@ -2,9 +2,9 @@ const { exec } = require('child_process');
 const fs = require('fs');
 const { expect } = require('chai');
 // Use package json as a view file because it contains a json and is perfect for this test
-const view = require('../../package.json');
+const view = require('../package.json');
 const path = require('path');
-const render = require('../../lib/render');
+const render = require('../lib/render');
 
 /* Note: to test if the binary script can be installed correctly, from the root of this repo call:
  * $ npm link
@@ -13,9 +13,9 @@ const render = require('../../lib/render');
  */
 
 describe('micromustache binary', () => {
-  const scriptPath = path.resolve(__dirname, '../../bin/micromustache');
-  const templatePath = path.resolve(__dirname, 'template.txt');
-  const viewPath = path.resolve(__dirname, '../../package.json');
+  const scriptPath = path.resolve(__dirname, 'micromustache');
+  const templatePath = path.resolve(__dirname, 'template.spec.txt');
+  const viewPath = path.resolve(__dirname, '../package.json');
 
   it('produces expected results for a valid template and view file', (done) => {
     const template = fs.readFileSync(templatePath, { encoding: 'utf8' });
@@ -29,28 +29,28 @@ describe('micromustache binary', () => {
     });
   });
 
-  it('fils if templatePath is missing', (done) => {
+  it('fails if templatePath is missing', (done) => {
     exec(`${scriptPath}`, (execErr) => {
       done();
       expect(execErr).not.to.be.null;
     });
   });
 
-  it('fils if templatePath is pointing to an invalid file', (done) => {
+  it('fails if templatePath is pointing to an invalid file', (done) => {
     exec(`${scriptPath} .`, (execErr) => {
       done();
       expect(execErr).not.to.be.null;
     });
   });
 
-  it('fils if viewPath is missing', (done) => {
+  it('fails if viewPath is missing', (done) => {
     exec(`${scriptPath} ${templatePath}`, (execErr) => {
       done();
       expect(execErr).not.to.be.null;
     });
   });
 
-  it('fils if viewPath is pointing to an invalid file', (done) => {
+  it('fails if viewPath is pointing to an invalid file', (done) => {
     exec(`${scriptPath} ${templatePath} .`, (execErr) => {
       done();
       expect(execErr).not.to.be.null;
