+++ bin/template.spec.txt
