+++ lib/compile.spec.js
@@ -1,7 +1,7 @@
 const { expect } = require('chai');
-const compile = require('../../lib/compile');
+const compile = require('./compile');
 
-describe('compile()', () => {
+describe('#compile()', () => {
   it('is a function', () => {
     expect(compile).to.be.a('function');
   });
