+++ lib/render.spec.js
@@ -1,5 +1,5 @@
 const { expect }= require('chai');
-const render = require('../../lib/render');
+const render = require('./render');
 
 describe('#render()', () => {
   it('is a function', () => {
