--- test/.eslintrc.json
@@ -1,17 +1,0 @@
-{
-  "env": {
-    "mocha": true
-  },
-  "globals": {
-    "expect": false
-  },
-  "plugins": [
-    "chai-expect"
-  ],
-  "rules": {
-    "chai-expect/no-inner-compare": ["error"],
-    "chai-expect/missing-assertion": ["error"],
-    "chai-expect/terminating-properties": ["error"],
-    "no-unused-expressions": "off"
-  }
-}
